document stripe and fix ecko plugins issues (#3)
