+++ ecko-plugins-stripe/ecko-plugins-stripe.gemspec
@@ -11,8 +11,9 @@ Gem::Specification.new do |spec|
   spec.summary = 'Mastodon stripe plugin, extendable to ecko-plugins sdk'
   spec.description = ''
   spec.homepage = 'https://github.com/magicstone-dev/ecko'
-  spec.license = 'GPL/AGPL'
+  spec.license = 'GPL-3.0+'
   spec.required_ruby_version = '>= 2.6.0'
+  spec.files = Dir['lib/**/*.rb', 'app/**/*.rb'].to_a
 
   spec.metadata['allowed_push_host'] = 'https://github.com/magicstone-dev/ecko'
 
@@ -30,10 +31,12 @@ Gem::Specification.new do |spec|
   spec.bindir = 'exe'
   spec.executables = spec.files.grep(%r{\Aexe/}) { |f| File.basename(f) }
   spec.require_paths = %w[lib app]
+  spec.add_dependency 'ecko-plugins', '~> 0.1.6'
   spec.add_dependency 'stripe', '~> 5.39'
 
   # Uncomment to register a new dependency of your gem
   # spec.add_dependency 'example-gem', '~> 1.0'
+  spec.add_dependency 'rails', '>= 6.1.4', '< 7.0.0'
 
   # For more information and examples about making a new gem, checkout our
   # guide at: https://bundler.io/guides/creating_gem.html
