+++ ecko-plugins-stripe/lib/ecko/plugins/stripe/configurations.rb
@@ -11,18 +11,24 @@ module Ecko
           @schema = schema
         end
 
+        # Stripe API key configured through schema
         def stripe_api_key
           schema[:stripe_api_key]
         end
 
+        # The state secret is used if we have to encrypt any state on callbacks
+        # Currently this hasn't been used.
         def state_secret
           schema[:state_secret]
         end
 
+        # Configured currency while checkout process.
         def currency
           schema[:currency] || 'USD'
         end
 
+        # Callback url which helps in returning to the correct url when a payment is complete or
+        # cancelled
         def callback
           schema[:callback]
         end
