+++ ecko-plugins-stripe/spec/ecko/plugins/stripe_spec.rb
@@ -1,11 +1,7 @@
 # frozen_string_literal: true
 
 RSpec.describe Ecko::Plugins::Stripe do
-  it "has a version number" do
+  it 'has a version number' do
     expect(Ecko::Plugins::Stripe::VERSION).not_to be nil
-  end
-
-  it "does something useful" do
-    expect(false).to eq(true)
   end
 end
