+++ ecko-plugins/.travis.yml
@@ -3,5 +3,5 @@ sudo: false
 language: ruby
 cache: bundler
 rvm:
-  - 2.6.3
+  - 2.7.4
 before_install: gem install bundler -v 1.17.2
