+++ ecko-plugins/lib/ecko/plugins/version.rb
@@ -1,5 +1,5 @@
 module Ecko
   module Plugins
-    VERSION = "0.1.1"
+    VERSION = '0.1.6'
   end
 end
